Fixing an NPE during undeploy/redeploy of .rar
Tests Run : QL, connector-dev


svn path=/trunk/; revision=27847


Former-commit-id: 71ca676973e1ed8f075cc7e889182659d36255fa

+++ v3/connectors/connectors-runtime/src/main/java/com/sun/enterprise/connectors/module/ConnectorDeployer.java
@@ -233,8 +233,8 @@ public class ConnectorDeployer extends JavaEEDeployer<ConnectorContainer, Connec
      */
     public void clean(DeploymentContext dc) {
         UndeployCommandParameters dcp = dc.getCommandParameters(UndeployCommandParameters.class);
-        if( dcp.origin == OpsParams.Origin.undeploy){
-            if(dcp.cascade){
+        if( dcp != null && dcp.origin == OpsParams.Origin.undeploy){
+            if(dcp.cascade != null && dcp.cascade){
                 deleteAllResources(dcp.name(), dcp.target);
             }
         }
